Create manager lib for StatsBootstrapAtomService
Manages binder lifecycle & provides interfaces for the underlying IPCs.
Callers can use StatsBootstrapAtomClient::reportBootstrapAtom. Under the
hood, we use checkService, which will not wait for the service. This is
so we avoid blocking the client. If StatsBootstrapAtomService is not
available, the log will be dropped, and false will be returned.

Test: manually verified atoms could be logged to statsd
Bug: 206644307
Change-Id: I63ae62f1ce4d1dc337f99c9e3b35b2496131c582